Revert "Fixed a basic Gameloop system"
This reverts commit f5f1b1a48eff55d2a9ffbd2e4929d04bd34a2b3f.
