+++ FranticFour/Assets/01_Scripts/Controlls/MovementController.cs
@@ -77,9 +77,8 @@ public class MovementController : MonoBehaviour
 
     private void MovePlayer()
     {
-        float xMovement = Input.GetAxis(controller.Horizontal);
-        float yMovement = Input.GetAxis(controller.Vertical);
-        Vector2 movement = new Vector2(xMovement, yMovement).normalized;
+        Vector2 movement = new Vector2(Input.GetAxis(controller.Horizontal), Input.GetAxis(controller.Vertical))
+            .normalized;
         rb2d.AddForce(movement * movementSpeed);
     }
 
@@ -91,6 +90,7 @@ public class MovementController : MonoBehaviour
         //yield return new WaitForSeconds(jumpCooldown);
         canJump = true;
     }
+
     private void StartJumping()
     {
         rb2d.freezeRotation = true;
@@ -102,6 +102,7 @@ public class MovementController : MonoBehaviour
             new Vector3(transform.localScale.x + 2, transform.localScale.y + 2,
                 5); // remove this once we have animation
     }
+
     private void EndJumping()
     {
         rb2d.freezeRotation = false;
@@ -144,6 +145,7 @@ public class MovementController : MonoBehaviour
     {
         StartCoroutine(HandlePush(pushForce));
     }
+
     IEnumerator HandlePush(Vector2 pushForce)
     {
         freezeInput = true;
@@ -157,6 +159,7 @@ public class MovementController : MonoBehaviour
     {
         StartCoroutine(HandleStun(duration));
     }
+
     IEnumerator HandleStun(float duration)
     {
         freezeInput = true;
