--- FranticFour/Assets/01_Scripts/GameLoopController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3fe43d1dba5a932459a100223a9fe85c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
