--- FranticFour/Assets/01_Scripts/PlayerActionsController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d1f4cb0f3da63fd46ab91c1b627c354b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
