+++ FranticFour/Assets/01_Scripts/PreyTrap.cs
@@ -7,8 +7,7 @@ public class PreyTrap : MonoBehaviour
     [SerializeField] float stunDelay = 2;
     private void OnTriggerEnter2D(Collider2D other)
     {
-        if (other.CompareTag("Player") && other.gameObject.layer != 9) 
-        {
+        if (other.CompareTag("Player") && other.gameObject.layer != 9) {
             GetComponent<BoxCollider2D>().enabled = false;
             other.GetComponent<MovementController>().GetStunned(stunDelay);
             Destroy(gameObject); // remove this once animation destroys the trap.
