+++ FranticFour/Assets/01_Scripts/PushController.cs
@@ -23,7 +23,6 @@ public class PushController : MonoBehaviour
             return;
         MovementController closestTarget = GetClosestTarget();
         closestTarget.GetPushed(pushForce);
-        closestTarget.GetComponent<Player>().GetPushedBy(GetComponentInParent<Player>());
     }
 
     private MovementController GetClosestTarget()
